Improve the UX of using huggingface models & embeddings
